+++ ThirdPartyAdapters/applovin/applovin/src/main/java/com/applovin/mediation/rtb/AppLovinRtbInterstitialRenderer.java
@@ -1,7 +1,10 @@
 package com.applovin.mediation.rtb;
 
+import static android.util.Log.DEBUG;
+import static android.util.Log.WARN;
+import static com.applovin.mediation.ApplovinAdapter.log;
+
 import android.content.Context;
-import android.util.Log;
 import com.applovin.adview.AppLovinInterstitialAd;
 import com.applovin.adview.AppLovinInterstitialAdDialog;
 import com.applovin.mediation.AppLovinUtils;
@@ -11,6 +14,7 @@ import com.applovin.sdk.AppLovinAdDisplayListener;
 import com.applovin.sdk.AppLovinAdLoadListener;
 import com.applovin.sdk.AppLovinAdVideoPlaybackListener;
 import com.applovin.sdk.AppLovinSdk;
+import com.google.android.gms.ads.AdError;
 import com.google.android.gms.ads.mediation.MediationAdLoadCallback;
 import com.google.android.gms.ads.mediation.MediationInterstitialAd;
 import com.google.android.gms.ads.mediation.MediationInterstitialAdCallback;
@@ -22,8 +26,6 @@ import com.google.android.gms.ads.mediation.MediationInterstitialAdConfiguration
 public final class AppLovinRtbInterstitialRenderer
     implements MediationInterstitialAd, AppLovinAdLoadListener, AppLovinAdDisplayListener,
     AppLovinAdClickListener, AppLovinAdVideoPlaybackListener {
-
-  private static final String TAG = AppLovinRtbInterstitialRenderer.class.getSimpleName();
 
   /**
    * Data used to render an RTB interstitial ad.
@@ -80,7 +82,7 @@ public final class AppLovinRtbInterstitialRenderer
   // region AppLovin Listeners
   @Override
   public void adReceived(AppLovinAd ad) {
-    Log.d(TAG, "Interstitial did load ad: " + ad.getAdIdNumber());
+    log(DEBUG, "Interstitial did load ad: " + ad.getAdIdNumber());
 
     this.ad = ad;
     mInterstitalAdCallback = callback.onSuccess(AppLovinRtbInterstitialRenderer.this);
@@ -88,40 +90,39 @@ public final class AppLovinRtbInterstitialRenderer
 
   @Override
   public void failedToReceiveAd(int code) {
-    Log.e(TAG, "Failed to load interstitial ad with error: " + code);
-
-    int admobErrorCode = AppLovinUtils.toAdMobErrorCode(code);
-    callback.onFailure(Integer.toString(admobErrorCode));
+    AdError error = AppLovinUtils.getAdError(code);
+    log(WARN, error.getMessage());
+    callback.onFailure(error);
   }
 
   @Override
   public void adDisplayed(AppLovinAd ad) {
-    Log.d(TAG, "Interstitial displayed.");
+    log(DEBUG, "Interstitial displayed.");
     mInterstitalAdCallback.reportAdImpression();
     mInterstitalAdCallback.onAdOpened();
   }
 
   @Override
   public void adHidden(AppLovinAd ad) {
-    Log.d(TAG, "Interstitial hidden.");
+    log(DEBUG, "Interstitial hidden.");
     mInterstitalAdCallback.onAdClosed();
   }
 
   @Override
   public void adClicked(AppLovinAd ad) {
-    Log.d(TAG, "Interstitial clicked.");
+    log(DEBUG, "Interstitial clicked.");
     mInterstitalAdCallback.reportAdClicked();
     mInterstitalAdCallback.onAdLeftApplication();
   }
 
   @Override
   public void videoPlaybackBegan(AppLovinAd ad) {
-    Log.d(TAG, "Interstitial video playback began.");
+    log(DEBUG, "Interstitial video playback began.");
   }
 
   @Override
   public void videoPlaybackEnded(AppLovinAd ad, double percentViewed, boolean fullyWatched) {
-    Log.d(TAG, "Interstitial video playback ended at playback percent: " + percentViewed + "%.");
+    log(DEBUG, "Interstitial video playback ended at playback percent: " + percentViewed + "%.");
   }
   // endregion